add test cases for History
