+++ tests/Feature/Backend/History/HistoryRenderTest.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Tests\Feature\Backend\History;
+
 use Tests\BrowserKitTestCase;
 
 /**
@@ -7,7 +9,8 @@ use Tests\BrowserKitTestCase;
  */
 class HistoryRenderTest extends BrowserKitTestCase
 {
-    public function testDashboardDisplaysHistory()
+    /** @test **/
+    public function admin_users_can_see_history_on_dashboard()
     {
         $this->actingAs($this->admin);
 
@@ -23,23 +26,8 @@ class HistoryRenderTest extends BrowserKitTestCase
              ->see('<strong>'.$this->admin->name.'</strong> '.trans('history.backend.users.created').$this->user->name);
     }
 
-    public function testTypeDisplaysHistory()
-    {
-        $this->actingAs($this->admin);
-
-        history()
-            ->withType('User')
-            ->withText(trans('history.backend.users.created').$this->user->name)
-            ->withEntity($this->user->id)
-            ->withIcon('plus')
-            ->withClass('bg-green')
-            ->log();
-
-        $this->visit('/admin/access/user')
-             ->see('<strong>'.$this->admin->name.'</strong> '.trans('history.backend.users.created').$this->user->name);
-    }
-
-    public function testEntityDisplaysHistory()
+    /** @test **/
+    public function admin_users_can_see_history_on_user_show_page()
     {
         $this->actingAs($this->admin);
